requirements.txt
streamlit
pillow
numpy
onnxruntime